+++ packages/ui/dropdown-menu/src/Checkbox/index.ts
@@ -1,0 +1,1 @@
+export * from './DropdownMenuCheckbox.js';