+++ packages/ui/dropdown-menu/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pikas-ui/dropdown-menu
 
+## 3.4.10
+
+### Patch Changes
+
+- fix and add auto loading
+- Updated dependencies
+  - @pikas-ui/menu@3.3.10
+
 ## 3.4.9
 
 ### Patch Changes